+++ RELEASE_NOTES.md
@@ -1,13 +1,13 @@
 ## VPinFE Release Notes
 
 ### Summary
-This is the offical release of vpinfe v1.1.41.
+This is the offical release of vpinfe v1.1.42
 
 ### What's New
-- More rom support/ ini support
+- "Last Played" now sorts on "LastRun".
+- Added sort by "LastRun" to collections.
 
 ### Fixes
 None
 
 ### Notes
-None